Extract invalid data alert
